requirements.txt
GitPython==3.1.44
Pillow==11.3.0
accelerate==1.10.1
bitsandbytes==0.48.0
clean-fid==0.1.35
diffusers==0.35.1
diskcache==5.6.3
einops==0.8.1
facexlib==0.3.0
fastapi==0.112.4
httpcore==0.15.0
httpx==0.24.1
huggingface-hub==0.34.1
inflection==0.5.1
joblib==1.5.1
jsonmerge==1.8.0
kornia==0.6.7
lark==1.2.2
lightning==2.5.1
loadimg==0.1.2
numpy==1.26.4
omegaconf==2.2.3
open-clip-torch==2.32.0
opencv-python==4.8.1.78
peft==0.17.1
piexif==1.1.3
pillow-heif==0.22.0
protobuf==4.25.7
psutil==5.9.8
pydantic-core==2.23.4
pydantic==2.9.2
resize-right==0.0.2
safetensors==0.6.2
scikit-image==0.25.2
setuptools==69.5.1
spandrel-extra-arches==0.2.0
spandrel==0.4.1
tomesd==0.1.3
torchdiffeq==0.2.5
torchsde==0.2.6
tqdm==4.67.1
transformers==4.56.2
torch
